requirements.txt
alembic==1.17.2
blinker==1.9.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
Flask==3.1.2
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.3.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
python-dotenv==1.2.1
requests==2.32.5
SQLAlchemy==2.0.45
typing_extensions==4.15.0
urllib3==2.6.2
Werkzeug==3.1.4
